Found the meaning of 'module'
No real methods yet, just playing with Class variables and constructors. I just found out, that a module is something handy. Without it, my method in user didn't work, because there seams to be a Ruby Class called Item as well and for that reason I had a parameter error.