Almost there
TradeItem and TradeItemTest are done now.
Item takes to responsebility of to whom it belongt, while the user takes responsibility about
his credit and payment.

Changed Arrays to Hash. Key is the name of the item and the item itself is of course the value.

Still missing are the tests of user for the major methods buy_item.

But not tonight. Sleep tight everone...

	modified:   app/Trade_Market/trade_item.rb
	modified:   app/Trade_Market/user.rb
	modified:   test/trade_item_test.rb
	modified:   test/user_test.rb